core, core/vm: moved logger and added gas cost to struct logging
